email-agent/backend/pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "Production-grade AI Email Agent — classifies, summarizes, and drafts email replies"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.150.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "cryptography>=43.0.0",
    "python-multipart>=0.0.12",
    "structlog>=24.0.0",
    "slowapi>=0.1.9",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pre-commit>=4.0.0",
    "bandit>=1.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101"]  # Allow assert in tests

[tool.ruff.format]
quote-style = "double"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
